third_party/blink/web_tests/virtual/xslt-disabled/fast/xsl/transformToFragment-XML-declaration-expected.txt

CONSOLE WARNING: XSLTProcessor and XSLT Processing Instructions have been removed in this browser. See https://chromestatus.com/feature/4709671889534976.
Test for bug 10419: XSLTProcessor transformToFragment fails because of an XML declaration.

FAILURE: NotSupportedError: Failed to construct 'XSLTProcessor': XSLT is disabled
